% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{allpairwise}
\alias{allpairwise}
\title{Title}
\usage{
allpairwise(xs, ys)
}
\arguments{
\item{xs}{}
}
\value{

}
\description{
Title
}
\examples{

n <- 6
H <- 0.578

lambda_quantiles <- lognorm_quantiles(meanlog = log(hatlambda), sdlog = H, n = n)
mu_quantiles <- lognorm_quantiles(meanlog = log(hatmu), sdlog = H, n = n)

allpairwise(lambda_quantiles, mu_quantiles)
}
