% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDS.R
\docType{package}
\name{BDS}
\alias{BDS}
\title{BDS: Birth death shift model}
\description{
This package implements the birth-death shift model for diversification
analysis based on reconstructed phylogenies.
}
\section{BDS functions}{

Functions here...
}

