% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_death_shift.R
\name{birth_death_shift}
\alias{birth_death_shift}
\title{Branch-specific speciation and extinction rate estimation}
\usage{
birth_death_shift(phy, lambda, mu, eta, rho, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Branch-specific speciation and extinction rate estimation
}
\examples{
library(Pesto)
library(ape)
library(ggtree)
library(tidytree)
library(ggplot2)
library(dplyr)

data(primates)

lambda <- c(0.1, 0.2, 0.3, 0.4, 0.20)
mu <- c(0.05, 0.15, 0.05, 0.15, 0.25)
rho <- 0.67
eta <- 0.008

td <- birth_death_shift(primates, lambda, mu, eta, rho)

th <- max(node.depth.edgelength(primates))
p1 <- ggtree(td, aes(color = `Speciation rate`)) +
             geom_tiplab(size = 3) +
             theme(legend.position = c(0.15, 0.8)) +
             xlim(c(0.0, th + 10))

plot(p1)
}
