% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_death_shift.R
\name{birth_death_shift2}
\alias{birth_death_shift2}
\title{Title}
\usage{
birth_death_shift2(phy, lambda, mu, eta, rho, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
\examples{
library(Pesto)
library(ape)
library(ggtree)
library(tidytree)
library(ggplot2)
library(dplyr)

data(primates)

lambda <- c(0.1, 0.2, 0.3, 0.4, 0.20)
mu <- c(0.05, 0.15, 0.05, 0.15, 0.25)
rho <- 0.67
eta <- 0.008

td <- birth_death_shift2(primates, lambda, mu, eta, rho)

th <- max(ape::node.depth.edgelength(primates))
p1 <- ggtree::ggtree(td, ggplot2::aes(color = `Speciation rate`)) +
             ggtree::geom_tiplab(size = 3) +
             ggplot2::theme(legend.position = c(0.15, 0.8)) +
             ggplot2::xlim(c(0.0, th + 10))

plot(p1)
}
