% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcPestoUtil.R
\name{CalcPestoUtil}
\alias{CalcPestoUtil}
\title{Intakes qualitative survey data}
\usage{
CalcPestoUtil(f_dat3 = pesto_exsumdf)
}
\arguments{
\item{f_dat3}{Output from SummarizePestoDF, default is to use pesto_exsumdf}
}
\value{
A data frame with utility for each package
}
\description{
Intakes qualitative survey data
}
