% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nshifts.R
\name{compute_nshifts}
\alias{compute_nshifts}
\title{Compute the expected number of shifts in rate category per branch}
\usage{
compute_nshifts(phy, lambda, mu, eta, rho, ntimeslices = 500)
}
\arguments{
\item{ntimeslices}{}
}
\value{

}
\description{
Compute the expected number of shifts in rate category per branch
}
\examples{
library(Pesto)
library(ggtree)
library(ggplot2)
data(primates)

lambda <- c(0.1, 0.2, 0.3, 0.4)
mu <- c(0.05, 0.15, 0.05, 0.15)
eta <- 0.008
rho <- 0.67

td <- compute_nshifts(primates, lambda, mu, eta, rho)

th <- max(ape::node.depth.edgelength(primates))
p1 <- ggtree(td, aes(color = nshifts)) +
             geom_tiplab(size = 3) +
             theme(legend.position = c(0.15, 0.8)) +
             xlim(c(0.0, th + 10)) +
             scale_colour_gradient(low = "black", high = "red")

plot(p1)
}
