% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesto_run.R
\name{.find_pestpp_exe}
\alias{.find_pestpp_exe}
\title{Find PEST++ executable}
\usage{
.find_pestpp_exe(name, user_path = NULL)
}
\arguments{
\item{name}{Name of the executable (e.g., "pestpp-ies")}

\item{user_path}{User-specified path (overrides search)}
}
\value{
Path to executable
}
\description{
Find PEST++ executable
}
\keyword{internal}
