% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_io.R
\name{.read_ensemble_binary}
\alias{.read_ensemble_binary}
\title{Internal binary ensemble reader}
\usage{
.read_ensemble_binary(file)
}
\arguments{
\item{file}{Path to binary file}
}
\value{
data.table
}
\description{
Internal binary ensemble reader
}
\keyword{internal}
