% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constantbdp.R
\name{estimate_constant_bdp}
\alias{estimate_constant_bdp}
\title{ML estimation of rates under the constant-rate Birth-Death process}
\usage{
estimate_constant_bdp(phy, rho = 1)
}
\arguments{
\item{phy}{an object of class phylo}

\item{rho}{the taxon sampling fraction. Must be above 0.0 and at maximum 1.0}
}
\value{

}
\description{
ML estimation of rates under the constant-rate Birth-Death process
}
\examples{
library(Pesto)

data(primates)
rho <- 0.67

res <- estimate_constant_bdp(primates, rho)
}
