% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extintion_probability.R
\name{extinction_probability}
\alias{extinction_probability}
\title{Compute the extinction probability E(t)}
\usage{
extinction_probability(phy, lambda, mu, rho)
}
\arguments{
\item{rho}{}
}
\value{

}
\description{
Compute the extinction probability E(t)
}
