% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstention.R
\name{is_pesto_abstention}
\alias{is_pesto_abstention}
\title{Is an object a PESTO abstention?}
\usage{
is_pesto_abstention(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
A single logical: \code{TRUE} when \code{x} is a \code{pesto_abstention}.
}
\description{
Is an object a PESTO abstention?
}
\examples{
is_pesto_abstention(pesto_abstention("over_determined"))
is_pesto_abstention(42)
}
