% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lognorm_quantiles}
\alias{lognorm_quantiles}
\title{lognormal quantiles}
\usage{
lognorm_quantiles(meanlog = log(0.2), sdlog = 0.587, n = 6)
}
\arguments{
\item{meanlog}{the mean of the normal distribution on a log scale}

\item{sdlog}{the standard deviation of the normal distribution on a log scale. sdlog = 0.587 represents a distribution whose 2.5-97.5\% quantile spans one order of magnitude}

\item{k}{the number of discrete quantiles to be returned}
}
\value{

}
\description{
lognormal quantiles
}
\examples{

q <- lognorm_quantiles(0.2, n = 8)
}
