% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePestoDF.R
\name{MakePestoDF}
\alias{MakePestoDF}
\title{Intakes qualitative survey data}
\usage{
MakePestoDF(f_dat = pesto_exinput)
}
\arguments{
\item{f_dat}{The elicited data, default is to use pesto_exinput}
}
\value{
A data frame with value dist for each category in each package
}
\description{
Intakes qualitative survey data
}
