% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pesto.R
\docType{package}
\name{Pesto}
\alias{Pesto}
\title{Pesto: Phylogenetic Estimation of Shifts in the Tempo of Origination}
\description{
This package implements the birth-death shift model for diversification
analysis based on reconstructed phylogenies.
}
\section{Pesto functions}{

Functions here...
}

