% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pesto_assessmentkey}
\alias{pesto_assessmentkey}
\title{More detailed description of the six performance categories}
\format{
A data frame with 6 rows and 3 variables:
\describe{
\item{assessment_category}{The short notation to refer to the performance category}
\item{assessment_question}{The question (more or less) posed in the survey to the user to assess the package's performance in the category}
\item{performance_metric}{A nice, concise wording for use in figures}
}
}
\usage{
pesto_assessmentkey
}
\description{
Includes the question (more or less) used in the survey
}
\keyword{datasets}
