% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pesto_exdf}
\alias{pesto_exdf}
\title{Example of output from MakePestoDF fxn}
\format{
A data frame with 90 rows and 8 variables:
\describe{
\item{title}{Title provided in pesto_exinput}
\item{package_title}{Same}
\item{weight}{Same}
\item{assessment_category}{Same}
\item{rating_text}{Same}
\item{confidence}{Same}
\item{score}{The rating text and confidence translated into a score for each value_bin}
\item{value_bin}{The value bin, corresponding to very low, low, med, high, very high performance}
}
}
\usage{
pesto_exdf
}
\description{
Used as default data in fxn
}
\keyword{datasets}
