% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pesto_exinput}
\alias{pesto_exinput}
\title{Example elicitation data (fair dust on lettuce)}
\format{
A data frame with an expanding number of rows and 4 variables:
\describe{
\item{title}{The user provided title for the comparison(s), should include the country, crop, and intervention(s)}
\item{package_title}{The user provided title for the package being evaluated}
\item{weight}{The weight assigned to each of the six impact categories, must add to 100}
\item{assessment_category}{The performance category}
\item{assignment_text}{The text assignment of the package performance given by the user to that category
(very low, low, medium, high, very high)}
\item{assignment_confidence}{The confidence the user has in the given assignment (L, M, H, VH)}
}
}
\usage{
pesto_exinput
}
\description{
An example of the elicitation data format to be input into the package function
}
\keyword{datasets}
