% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesto_run.R
\name{pesto_glm}
\alias{pesto_glm}
\title{Run PEST++ GLM (Gauss-Levenberg-Marquardt)}
\usage{
pesto_glm(
  pst_file,
  exe = NULL,
  noptmax = 20,
  extra_args = list(),
  working_dir = NULL,
  verbose = TRUE
)
}
\arguments{
\item{pst_file}{Character. Path to the .pst control file.}

\item{exe}{Character. Path to pestpp-glm executable.}

\item{noptmax}{Integer. Maximum number of iterations.}

\item{extra_args}{Named list. Additional PEST++ arguments.}

\item{working_dir}{Character. Working directory.}

\item{verbose}{Logical. Print output.}
}
\value{
A list of class \code{pesto_glm_result}.
}
\description{
Executes the pestpp-glm algorithm for deterministic parameter estimation.
}
