% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig-theme.R
\docType{data}
\name{pesto_th1}
\alias{pesto_th1}
\title{Testy function #1}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 11.
}
\usage{
pesto_th1
}
\value{
a theme
}
\description{
Testy function #1
}
\keyword{datasets}
