% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_identifiability}
\alias{plot_identifiability}
\title{Plot Parameter Identifiability}
\usage{
plot_identifiability(
  jco_file = NULL,
  pst = NULL,
  n_sv = NULL,
  title = "Parameter Identifiability"
)
}
\arguments{
\item{jco_file}{Character. Path to .jco (Jacobian) file.}

\item{pst}{A \code{pesto_pst} object for parameter names.}

\item{n_sv}{Integer. Number of singular values to retain.}

\item{title}{Character. Plot title.}
}
\value{
A ggplot2 object.
}
\description{
Creates a bar plot of parameter identifiability based on
the singular value decomposition of the Jacobian matrix.
}
