% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_rff_surrogate}
\alias{predict_rff_surrogate}
\title{Predict with RFF Sparse GP Surrogate}
\usage{
predict_rff_surrogate(rff, X_new)
}
\arguments{
\item{rff}{A trained RFF model (from \code{train_rff_surrogate}).}

\item{X_new}{Matrix (m x npar). New parameter sets.}
}
\value{
A list with mean predictions and approximate uncertainties.
}
\description{
Predict with RFF Sparse GP Surrogate
}
