% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst_io.R
\name{read_pst}
\alias{read_pst}
\title{Read a PEST Control File (.pst)}
\usage{
read_pst(file)
}
\arguments{
\item{file}{Character. Path to the .pst file.}
}
\value{
A list of class \code{pesto_pst} containing:
\describe{
\item{control_data}{Control section parameters (NPAR, NOBS, etc.)}
\item{parameter_groups}{data.table of parameter group definitions}
\item{parameters}{data.table of parameter data}
\item{observation_groups}{data.table of observation group definitions}
\item{observations}{data.table of observation data}
\item{model_command}{Character vector of model command lines}
\item{template_files}{data.table of template/model input file pairs}
\item{instruction_files}{data.table of instruction/model output file pairs}
\item{prior_information}{data.table of prior information (if present)}
\item{pestpp_options}{Named list of ++ options}
}
}
\description{
Parses a PEST/PEST++ control file and returns a structured list
containing all sections: control data, parameter groups, parameter
data, observation groups, observation data, model command, and
template/instruction file pairs.
}
\examples{
pars <- data.table::data.table(
  parnme = c("k1", "k2", "k3"),
  partrans = c("log", "log", "none"),
  parchglim = "factor",
  parval1 = c(1.0, 0.5, 0.1),
  parlbnd = c(0.01, 0.001, 0.0),
  parubnd = c(100, 50, 1.0),
  pargp = c("hydraulic", "hydraulic", "storage")
)
obs <- data.table::data.table(
  obsnme = c("h1", "h2", "h3"),
  obsval = c(1.0, 2.0, 1.5),
  weight = c(1.0, 1.0, 0.5),
  obgnme = "head"
)
pst <- create_pest_scenario(pars, obs, model_command = "echo run")
tf <- tempfile(fileext = ".pst")
on.exit(unlink(tf), add = TRUE)
write_pst(pst, tf)
pst_back <- read_pst(tf)
pst_back$control_data$npar
pst_back$control_data$nobs
}
\seealso{
\code{\link[=write_pst]{write_pst()}}, \code{\link[=create_pest_scenario]{create_pest_scenario()}}
}
