% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizePestoDF.R
\name{SummarizePestoDF}
\alias{SummarizePestoDF}
\title{Intakes qualitative survey data}
\usage{
SummarizePestoDF(f_dat2 = pesto_exdf)
}
\arguments{
\item{f_dat2}{Output from MakePestoDF, default is to use pesto_exdf}
}
\value{
A data frame with value dist for each category in each package
}
\description{
Intakes qualitative survey data
}
