% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{treeprecompute}
\alias{treeprecompute}
\title{Precompute tree statistics}
\usage{
treeprecompute(phy)
}
\arguments{
\item{phy}{}
}
\value{

}
\description{
Add some tree statistics to the tree object, compute the branch traversal order (postorder).
}
\examples{
library(Pesto)

data(bears)

bears <- treeprecompute(bears)
}
