% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{allpairwise}
\alias{allpairwise}
\title{All pairwise combinations among x and y}
\usage{
allpairwise(xs, ys)
}
\arguments{
\item{xs}{}
}
\value{
All the pairwise combinations
}
\description{
All pairwise combinations among x and y
}
\examples{
n <- 6
H <- 0.578

lambda_quantiles <- lognorm_quantiles(meanlog = log(0.28), sdlog = H, n = n)
mu_quantiles <- lognorm_quantiles(meanlog = log(0.20), sdlog = H, n = n)

allpairwise(lambda_quantiles, mu_quantiles)
}
