% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesto-package.R
\docType{package}
\name{PESTO-package}
\alias{PESTO}
\alias{PESTO-package}
\title{PESTO: Parameter Estimation, Surrogates, and Tooling for Optimisation}
\description{
High-performance parameter estimation, uncertainty quantification, and inverse modelling toolkit built on modernised PEST++ algorithms. Provides native R interfaces to iterative ensemble smoothers (IES), Gauss-Levenberg-Marquardt (GLM) solvers, global sensitivity analysis, and multi-objective optimisation under uncertainty. Includes surrogate-accelerated IES via Gaussian Process and Random Fourier Features, adaptive SVD backends (randomised SVD, LAPACK, Eigen), and convergence-aware adaptive ensemble sizing. Designed for large-scale environmental, hydrological, and agricultural models with support for highly-parameterised problems (>100,000 parameters).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AAGI-AUS/PESTO}
  \item \url{https://aagi-aus.github.io/PESTO}
  \item Report bugs at \url{https://github.com/AAGI-AUS/PESTO/issues}
}

}
\author{
\strong{Maintainer}: Max Moldovan \email{max.moldovan@adelaide.edu.au} (\href{https://orcid.org/0000-0001-9680-8474}{ORCID}) [copyright holder]

}
\keyword{internal}
