% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesto_run.R
\name{pesto_sweep}
\alias{pesto_sweep}
\title{Run PEST++ SWP (Parametric Sweep)}
\usage{
pesto_sweep(
  pst_file,
  par_ensemble,
  exe = NULL,
  working_dir = NULL,
  verbose = TRUE
)
}
\arguments{
\item{pst_file}{Character. Path to the .pst control file.}

\item{par_ensemble}{data.table or path. Parameter ensemble.}

\item{exe}{Character. Path to pestpp-swp executable.}

\item{working_dir}{Character. Working directory.}

\item{verbose}{Logical. Print output.}
}
\value{
A list containing observation outputs for each realisation.
}
\description{
Executes pestpp-swp for embarrassingly parallel model runs across
a parameter ensemble.
}
