% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst_io.R
\name{print.pesto_pst}
\alias{print.pesto_pst}
\title{Print method for pesto_pst objects}
\usage{
\method{print}{pesto_pst}(x, ...)
}
\arguments{
\item{x}{A \code{pesto_pst} object.}

\item{...}{Ignored.}
}
\description{
Print method for pesto_pst objects
}
