% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst_io.R
\name{read_pst}
\alias{read_pst}
\title{Read a PEST Control File (.pst)}
\usage{
read_pst(file)
}
\arguments{
\item{file}{Character. Path to the .pst file.}
}
\value{
A list of class \code{pesto_pst} containing:
\describe{
\item{control_data}{Control section parameters (NPAR, NOBS, etc.)}
\item{parameter_groups}{data.table of parameter group definitions}
\item{parameters}{data.table of parameter data}
\item{observation_groups}{data.table of observation group definitions}
\item{observations}{data.table of observation data}
\item{model_command}{Character vector of model command lines}
\item{template_files}{data.table of template/model input file pairs}
\item{instruction_files}{data.table of instruction/model output file pairs}
\item{prior_information}{data.table of prior information (if present)}
\item{pestpp_options}{Named list of ++ options}
}
}
\description{
Parses a PEST/PEST++ control file and returns a structured list
containing all sections: control data, parameter groups, parameter
data, observation groups, observation data, model command, and
template/instruction file pairs.
}
\examples{
# pst <- read_pst("model.pst")
# print(pst)
}
