% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl_wrapper.R
\name{treeprecompute}
\alias{treeprecompute}
\title{Title}
\usage{
treeprecompute(phy)
}
\arguments{
\item{phy}{}
}
\value{

}
\description{
Title
}
\examples{
library(BDS)

data(bears)

bears <- treeprecompute(bears)
}
