% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst_io.R
\name{write_pst}
\alias{write_pst}
\title{Write a PEST Control File (.pst)}
\usage{
write_pst(pst, file)
}
\arguments{
\item{pst}{A \code{pesto_pst} object (as returned by \code{\link[=read_pst]{read_pst()}}).}

\item{file}{Character. Output file path.}
}
\value{
Invisible \code{NULL}. File is written as a side effect.
}
\description{
Writes a \code{pesto_pst} object to a PEST-format control file.
}
